More codebase cleaning
