+++ src/main/java/com/xray/common/proxy/ClientProxy.java
@@ -1,8 +1,9 @@
 // Forge proxy for the client side.
 package com.xray.common.proxy;
 
-import com.xray.client.KeyBindingHandler;
+import com.xray.client.keybinding.InputEvent;
 import com.xray.client.gui.GuiOverlay;
+import com.xray.client.keybinding.KeyBindings;
 import com.xray.client.xray.XrayController;
 import com.xray.client.xray.XrayEventHandler;
 import com.xray.common.XRay;
@@ -25,10 +26,10 @@ public class ClientProxy
 	public void preInit(FMLPreInitializationEvent event) {
 		ConfigHandler.init( event.getSuggestedConfigurationFile() );
 
-		// Setup the keybindings
-		KeyBindingHandler.setup();
+		// Setup the keyBindings
+		KeyBindings.setup();
 
-		MinecraftForge.EVENT_BUS.register( new KeyBindingHandler() );
+		MinecraftForge.EVENT_BUS.register( new InputEvent() );
 		MinecraftForge.EVENT_BUS.register( new XrayEventHandler() );
 		MinecraftForge.EVENT_BUS.register( new GuiOverlay() );
 	}
